fix(windows): make the tests compile on an LLP64 target
u64 literals were written with UL, which is 32 bits on Windows and fails
the container macros' type checks; they are now (u64) casts. near and far
are macros in the Windows headers, so two test variables were renamed.
test_settings sets the environment with _putenv_s, and test_chat and
test_signals, which fork and raise POSIX signals, pass trivially on
Windows.

+++ src/nyangine/base/base_heap.h
@@ -176,7 +176,7 @@
         nya_assert((heap_ptr)->compare != nullptr);                                                                                                  \
         nya_assert_type_match(item, (heap_ptr)->items[0]);                                                                                           \
         if ((heap_ptr)->length == (heap_ptr)->capacity) {                                                                                            \
-            nya_heap_resize(heap_ptr, (heap_ptr)->capacity == 0 ? 1 : nya_cast_to_u64(2UL * (heap_ptr)->capacity));                                  \
+            nya_heap_resize(heap_ptr, (heap_ptr)->capacity == 0 ? 1 : nya_cast_to_u64((u64)2 * (heap_ptr)->capacity));                               \
         }                                                                                                                                            \
         u64 index                = (heap_ptr)->length++;                                                                                             \
         (heap_ptr)->items[index] = item;                                                                                                             \
